pyproject.toml
[tool.poetry]
name = "zeroshot"
version = "0.1.10"
description = "Image classifier with zero-shot learning."
license = "MIT"

authors = [
    "Nate Harada <hello@usezeroshot.com>"
]
repository = "https://github.com/moonshinelabs/zeroshot-python"
homepage = "https://github.com/moonshinelabs/zeroshot-python"

readme = ["README.md"]
keywords = ["zeroshot", "classifier", "cv"]

# Setup dependencies.
[tool.poetry.dependencies]
python = ">=3.10,<3.13"
numpy = "^1.26.0"
pillow = "^10.0.1"
types-pillow = "^10.0.0.3"
mypy = "^1.6.0"
onnxruntime = "^1.16.1"
requests = "^2.31.0"
torch = {version = "^2.1.0", optional = true, source = "pytorch"}
xformers = {version = "^0.0.22.post7", optional = true, source = "pytorch"}

[[tool.poetry.source]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
priority = "explicit"

[tool.poetry.extras]
torch = ["torch", "xformers"]

[tool.poetry.group.docs]
optional = true
[tool.poetry.group.docs.dependencies]
Sphinx = "^5.1.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
parameterized = "^0.9.0"

# MyPy is missing stubs, which is annoying
scikit-learn = "^1.3.2"
requests-mock = "^1.11.0"
[[tool.mypy.overrides]]
module = [
    "parameterized",
    "pillow_heif",
    "requests",
    "onnxruntime"
]
ignore_missing_imports = true
